# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "SMRITI-OS - Sovereign Retail OS",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/masks/entry": {
            "get": {
                "tags": [
                    "masks"
                ],
                "summary": "Get Entry Mask",
                "description": "Returns the field entry mask for the billing grid.\nDerived from Shoper9's AcceptDisplayDtls.",
                "operationId": "get_entry_mask_api_v1_masks_entry_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "trn_type",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Trn Type"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/masks/validation": {
            "get": {
                "tags": [
                    "masks"
                ],
                "summary": "Get Validation Mask",
                "description": "Returns business validation rules for the transaction.",
                "operationId": "get_validation_mask_api_v1_masks_validation_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "trn_type",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Trn Type"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/masks/hotkeys": {
            "get": {
                "tags": [
                    "masks"
                ],
                "summary": "Get Hotkey Mask",
                "description": "Returns institutional hotkey registry.",
                "operationId": "get_hotkey_mask_api_v1_masks_hotkeys_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "trn_type",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Trn Type"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/masks/zone_c": {
            "get": {
                "tags": [
                    "masks"
                ],
                "summary": "Get Zone C Mask",
                "description": "Returns Zone C (Right Panel) layout.",
                "operationId": "get_zone_c_mask_api_v1_masks_zone_c_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "trn_type",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Trn Type"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/masks/permissions": {
            "get": {
                "tags": [
                    "masks"
                ],
                "summary": "Get Permission Mask",
                "description": "Returns sensitive action permissions.",
                "operationId": "get_permission_mask_api_v1_masks_permissions_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "trn_type",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Trn Type"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}